Move finished views to separate packages
